pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[project]
name = "openenv-laptop-audio-ns"
version = "0.1.0"
description = "Adaptive noise suppression environment for Meta OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]==0.2.1",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "numpy",
    "soundfile",
    "gradio",
    "openai",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
packages = ["server", "env", "agent"]
